Small Optimisations

+++ source/tweener/EasingFuncs.cpp
@@ -3,37 +3,37 @@
 namespace liquid {
 namespace tweener {
 
-    double EasingFuncs::linear(double t)
+    float EasingFuncs::linear(float t)
     {
         return t;
     }
 
-    double EasingFuncs::sine(double t)
+    float EasingFuncs::sine(float t)
     {
-        static const double pi = 3.141592653589793238462;
+        static const float pi = 3.141592653589793238462;
         return 1 - std::cos(t * pi / 2.0);
     }
 
-    double EasingFuncs::quadratic(double t)
+    float EasingFuncs::quadratic(float t)
     {
         return std::pow(t, 2);
     }
 
-    double EasingFuncs::cubic(double t)
+    float EasingFuncs::cubic(float t)
     {
         return std::pow(t, 3);
     }
 
-    double EasingFuncs::elastic(double t)
+    float EasingFuncs::elastic(float t)
     {
-        static const double pi = 3.141592653589793238462;
-        const double v = t - 1;
-        const double p = 0.3;
+        static const float pi = 3.141592653589793238462;
+        const float v = t - 1;
+        const float p = 0.3;
 
         return -std::pow(2, 10 * v) * std::sin((v - p / 4.0) * 2.0 * pi / p);
     }
 
-    double EasingFuncs::exponential(double t)
+    float EasingFuncs::exponential(float t)
     {
         if (t == 0)
             return 0;
@@ -41,11 +41,11 @@ namespace tweener {
             return std::pow(2, 10 * (t - 1));
     }
 
-    double EasingFuncs::bounce(double t)
+    float EasingFuncs::bounce(float t)
     {
-        const double v = 1 - t;
-        double c;
-        double d;
+        const float v = 1 - t;
+        float c;
+        float d;
 
         if (v < (1.0 / 2.75))
         {
