+++ source/tweener/EasingFuncs.h
@@ -25,7 +25,7 @@ public:
       *
       * where f(t) = t
       */
-    static double linear(double t);
+    static float linear(float t);
 
     /** \brief Sine Easing
       * \param t Time-slice (0 - 1)
@@ -33,7 +33,7 @@ public:
       *
       * where f(t) = 1 - cos(t * pi / 2)
       */
-    static double sine(double t);
+    static float sine(float t);
 
     /** \brief Quadratic Easing
       * \param t Time-slice (0 - 1)
@@ -41,7 +41,7 @@ public:
       *
       * where f(t) = t^2
       */
-    static double quadratic(double t);
+    static float quadratic(float t);
 
     /** \brief Cubic Easing
       * \param t Time-slice (0 - 1)
@@ -49,7 +49,7 @@ public:
       *
       * where f(t) = t^3
       */
-    static double cubic(double t);
+    static float cubic(float t);
 
     /** \brief Elastic Easing
       * \param t Time-slice (0 - 1)
@@ -57,7 +57,7 @@ public:
       *
       * where f(t) = -2^(10 * (t - 1)) * sin(((t - 1) - 0.3 / 4) * 2 * pi / 0.3)
       */
-    static double elastic(double t);
+    static float elastic(float t);
 
     /** \brief Exponential Easing
       * \param t Time-slice (0 - 1)
@@ -65,7 +65,7 @@ public:
       *
       * where f(t) = 0 if t == 0 || 2^(10 * (t - 1))
       */
-    static double exponential(double t);
+    static float exponential(float t);
 
     /** \brief Bounce Easing
       * \param t Time-slice (0 - 1)
@@ -77,7 +77,7 @@ public:
       *              if (v < (2.5 / 2.75)) -> 1 - (7.5625 * (v - 2.25 / 2.75) * (v - 2.25 / 2.75) + 0.9375      or
       *              otherwise             -> 1 - (7.5625 * (v - 2.625 / 2.75) * (v - 2.625 / 2.75) + 0.984375
       */
-    static double bounce(double t);
+    static float bounce(float t);
 };
 
 #endif // _EASINGFUNCS_H
